Make sure that NewElems never need to deal with KindOfUninit bases
There was a strange case in LvalNew that was producing
KindOfUninits inside of arrays, and the lvalBlackHole was KindOfUninit
instead of KindOfNull.  Looking around, nothing appears to be using
the fact that these are Uninit, but I noticed Uninit (and KindOfRef)
handling in IncDecBody and decided to clean it up as part of this.
Also this removes the unused Variant::lvalInvalid.

Reviewed By: @bertmaher

Differential Revision: D1235606