Completely break Simplifier's dependency on IRBuilder
This makes Simplifier a standalone module, so it's now in shape to be
unit-tested sanely. The last thing to deal with was the calls to
IRBuilder::typeMightRelax, which I've replaced with a std::function.
The constructor now takes only IRUnit.

About the std::function: if I'm reading the implementation correctly (no
guarantees; reading libstdc++ source is no fun) then for sufficiently
small callable objects, no heap allocation happens. (This is certainly
the case in libc++). The sizeof the lambda I pass is 8, so if libstdc++
implements that optimization, it will definitely kick in. I considered
making std::function use smart::Allocator, but it turns out libstdc++
has not implemented that feature yet.

Reviewed By: @ottoni

Differential Revision: D1231776