Audit all ArrayInit call sites, change them to specify expected shape
I'm going to change ArrayInit to start with kMixedKind
instead of kPackedKind, since almost all callsites are going to
immediately insert a string.  Many packedToMixed calls seem to result
from this stuff in extension code right now, and this will also let us
devirtualize the calls in ArrayInit if we want.  The ones that weren't
actually Map/Struct-like I mostly ported to PackedArrayInit, except a
few cases that have mixed or integer keys, which I just used a
type-alias to flag as Mixed for now.

Reviewed By: @bertmaher

Differential Revision: D1240419