Reinstate full conversion, disallow \primitive
Now that D1229462 is ready, we don't need this any more. This
might break a little bit of Hack code that was written externally, so
give a reasonable error about it. (I don't feel too bad about breaking
backwards compat with the folks who have been using our nightlies --
there hasn't been a release yet; this will all be in 3.0.)

Closes #2135

Reviewed By: @gabelevi

Differential Revision: D1234791