EZC: fix bug in _zval_copy_ctor_func()
The type of the dest zval was copied from the source, but in the
KindOfStaticString case, the data of the dest zval is overwritten with a
plain string. This causes an assertion when the dest zval is used.

Enable the test which previously failed due to this bug.
Closes #2045

Reviewed By: @fredemmott

Differential Revision: D1240866

Pulled By: @ptarjan