Add control-flow arcs to RegionDesc
To support more general regions, it's necessary to specify which
control-flow arcs are within a region.  This diff adds basic support
for storing these arcs in the RegionDesc and changes the various
region selectors to store the arcs in the regions they produce.
Follow-up diffs will change the region translator to actually follow
what arcs are within the region.

Reviewed By: @bertmaher

Differential Revision: D1242572