Fix constant branch elimination
Constant branch elimination wasn't working with bc control
flow because, if the target of a constant branch was also the target
of some prior branch, then startBlock() would try to add a fallthrough
block to the (now eliminated) branch.  This diff catches that case and
instead adds an explicit fallthrough jump during startBlock(), which
can be removed during jumpopts if the block layout allows.

Simple example exhibiting the problem:

     JmpNZ B
  A: True
     JmpNZ B
  B: ...

Reviewed By: @oyamauchi

Differential Revision: D1243443