Fix a bug in JIT'd FPushClsMethodD in repo mode
If you are in an instance method, calling to an instance
method in a subclass with static syntax, we need to pass the $this
pointer along.  We diverge from zend in that we only carry $this along
if the current instance is a subclass of the subclass you are
dispatching to, so it's easiest to punt this for now to get it right.

Reviewed By: @elgenie

Differential Revision: D1245122