Break a bytecode dependency in the refcount optimizer
This is with the aim of unit-testing this component. In general, I think
knowledge like this (is it in a generator?) that we're gaining from
looking at bytecode should be encoded in the IR if it's necessary.

Anyway, what this code is actually interested in checking is (a) is it
the last instruction in an exit block; (b) is it NOT a RetCtrl from a
non-generator; (c) is it NOT in the catch block of a non-generator's
FunctionExitSurpriseHook. All this information is already there in the
IR.

Reviewed By: @bertmaher

Differential Revision: D1246382