Fix bugs related to old style constructors
Consider:

  class base {
    public function base() { ... }
  }

  clase derived extends base {
    public function base() { ... }
  }

  $d = new derived;

This would emit a FPushCtorD. When it got to annotate(), it would
annotate the push with the class name of the object ("derived") which
would end up calling "derived::base" instead of "base::base".

HHBBC had a similar bug where it assumed that a resolved ctor would
match the resolved class, making an FCallD "derived" "base".

Reviewed By: @markw65

Differential Revision: D1244171