Move static properties into RDS
Rather than maintaining an RDS handle to a pointer to an array of
TypedValue*'s and writing those pointers into a target cache, instead we
allocate the TypedValue's directly in RDS, and maintain an array of
handles.  This lets us obtain pointers to static properties without any
loads.

We replace LdClsPropAddrCached with LdClsPropAddrKnown, which is
codegen'd together with InitSProps.

Additionally, we eliminate the special-cased branch-or-catch logic for
LdClsPropAddr*, which was needed for Isset/Empty versus other S-opcodes.
Instead, we split LdClsPropAddr into LdClsPropAddrOrNull and
LdClsPropAddrOrRaise, and use CheckNonNull to branch.

This diff also eliminates LdClsStaticInitData (since there is no longer
an init vector for static props), as well as the SPropCache.

In addition, there are two bugfixes:

- Class::initSProps() can potentially throw, so we actually need catch
  blocks whenever we call Class::getSProp()---we now pass these in for
  IssetS and EmptyS along with the other S opcodes.

- We can only use the cached version of LdClsPropAddr when no 86sinit's
  will be called when we invoke getSProp().  Previously, we were only
  checking for such initializers on the Class in question, but not its
  inheritance chain, whose inits would also be called.

Reviewed By: @jdelong

Differential Revision: D1256221