Simplify VarEnv
Simplify previously overly complicated VarEnv::{attach,detach}() logic,
which is now possible as VarEnv is the only user of NameValueTable.
Move attached ActRec* to the NameValueTable, which now uses
KindOfNamedLocal instead of KindOfIndirect.

Makes it possible to migrate VarEnv from eagerly executed async function
or continuation to the ActRec inside the Continuation object (coming
soon in a follow-up diff).

Reviewed By: @jdelong

Differential Revision: D1250216