ProxyArray needs to call escalateForSort on its inner arrays before sorting
zend_hash_sort is currently just calling sort directly on the
array, which means it has no chance to escalate.  It shouldn't call
escalateForSort on the proxy array itself (because we need to keep
keep the Hashtable* constant), so proxy array needs to do it on its
inner array any time you tell it to sort.

Reviewed By: @edwinsmith

Differential Revision: D1248308