Cleanup in array-iterator helpers (mostly dead code removal)
iter_next wasn't used anymore, and iter_next_key is only
instantiated when withRef is true.  getContainerIter was also dead.
Removed some unnecessary casts in the function pointer table, and the
table didn't need to be exported in the header.  While here I also
figured we can directly send the iterNextKObject case to
iter_next_cold instead of to a function that just jumps to it
(probably not a common enough case to matter).

Reviewed By: @dariorussi

Differential Revision: D1248545