Fix second-pass CSE
During reoptimize we remove instructions from their block
before optimizing and re-appending them, so it always fails when we
CSE checks to see if an intruction's block is dominated by an
available expression.

The simplest solution seemed to be passing in the instruction's source
block along with the dominator list.

Reviewed By: @swtaarrs

Differential Revision: D1250860