HHVM extension: Fix UTF-8 error recovery in json_encode.
When reading a multi-byte sequence, if the utf-8 reader encounters an invalid byte,
abort the sequence, output U+FFFD, and restart scanning at the invalid byte in JSON_FB_LOOSE mode.

Overriding the unit tests as they are wrong after this fix. I have a separate diff D1201123 for the unit tests.

Reviewed By: @ealf

Differential Revision: D1172887