Retranslate without control flow if refcount opts fails
Merging FrameStacks in refcount opts is apparently pretty
tricky.  To keep making forward progress let's punt on that case for
now and try retranslating the region without control flow.  Even when
this particular case is fixed it still seems useful to have a bail-out
path to weaken optimizations and try again.

Reviewed By: @swtaarrs

Differential Revision: D1250987