Unit test for refcount optimizer
The solution was obvious all along: pass a FrameState into
optimizeRefcounts. I like this because it's putting a clue as to the
scope of optimizeRefcounts' work into its interface. It furthers my goal
of separating IRUnit from BCMarker, and thus my higher-level goal of
separating IR from bytecode.

Right now the refcount optimizer's unit test is trivial (can it optimize
IncRef; IncRef; DecRef; DecRef of the same thing?) because I want to
leave the writing of nontrivial tests for a diff without non-test
changes.

Reviewed By: @swtaarrs

Differential Revision: D1251523