Fix assertion failure with wrong type param arity
We wanted to assert that, when using a class, its methods
have the right arity. However, that check currently happens during the
typing phase, and so depending on ordering there's a chance that some
callsite might consume the return type with a wrong arity before we can
check that it's correct. Furthermore this check can't be moved up into
declaration since we need to get the typeclass of the target class in
order to determine what the arity should be in the first place. (I
suppose we could use the nast for this, but that sounds annoying.) So we
could define a new phase where this is checked and where we don't go
into typing if it fails... or we could just remove the asserts and let
typing deal with spewing the type errors later. We already are at least
intended to do something reasonable on arity mismatch since we omit this
check in silent mode anyways.

Reviewed By: @gabelevi

Differential Revision: D1246487