Builtin classes are always persistent
While debugging other issues, I noticed that we were
reading the Class* for Exception from the rds, even though
the handle was clearly a persistent handle.

Found that we were guarding on RepoAuthoritative - but thats
clearly too late (if the Class isn't persistent, for whatever
reason, we should not have allocated a persistent handle).

Reviewed By: @jdelong

Differential Revision: D1253986