Fix incompatible interface implementation error message
This one was a mess. We had a bunch of `Pos.none` lying around,
leading to an incredibly bad error message for the new test case in this
diff (distilled from a case that @KendallHopkins actually hit in real
code). This diff tries to give better witnesses in a lot of cases,
particularly fixing the error message in this case. It still isn't a
**great** error, but it at least is workable now, pointing to the
problem instead of into the ether :)

Reviewed By: @pikatchu

Differential Revision: D1252761