Fix leak when FPushCufArray throws
The object got written to the stack by the c++ helper,
but the catch trace also wrote the callable to the same stack slot.
This changes the c++ helper to write the object to the Cell with the
highest address in the ActRec, instead of the lowest, and also to make
the marker and the SyncPoint agree on where the stack is supposed to
be.

Reviewed By: @jdelong

Differential Revision: D1255189