Fix issues in makeExitImpl
It was being passed a vector of spillValues, which
was effectively a replacement evalStack. But the evalStack
was still being accessed implicitly by various things.

Swap the passed in spillValues with the evalStack (and swap
back at the end).

Also, we didn't properly update state in exit traces, resulting
in inconsistencies between the markers and the actual stack depth.

Reviewed By: @jdelong

Differential Revision: D1259246