Corrected the implementation of ReflectionParameter::isOptional
The old implementation of ReflectionParameter::isOptional only checked in there is a default value for the param, however that behavior is wrong.
This corrects the behavior to suit php5.5's implementation which dictates the param is optional only if it's at the end or is followed by an optional param.

Reviewed By: @ptarjan

Differential Revision: D1254654