Add a new admin command to dump the meta info out of APC
Often times the exact contents of the keys in APC isn't as useful as the sizes of the values, whether the item is in memory or in the archive, and the TTL. This adds a new admin command that can dump the meta information about APC in a space separated file. This can then easily be parsed by a variety of tools to do offline analysis. The short term goal is to run this on a few webis infrequently to dump the conents of apc to scuba so we can get a view of where the problems in APC are. A similar approach was done for memcache through the memcached composition analysis tables. This resulting investigation resulted in large efficiency improvements in the Cache tier. Note that the unit tests seem to be peremently postponed so I am marking this diff as override-unit-failures.

Reviewed By: @dariorussi

Differential Revision: D1257692