Fix stack spilling issues
peekSpillValues returned an array that was in the reverse
order of the array in m_evalStack. This is surprising, and in a few
places people had then manipulated it incorrectly (pushing a value on
the end, thinking that was the new "top"), resulting in incorrect
results when the stack was not empty.

This reverses the order of the elements in peekSpillValues, and fixes
the places that needed fixing (most did not, because they were backwards
to start with).

Reviewed By: @alexmalyshev

Differential Revision: D1259234