set default HTTP_VERSION in fastcgi transport
Flushing the output buffer before the end of request requires a
HTTP_VERSION of 1.1 in HHVM. FastCGI will currently set this to an empty
string when it is not specified in webserver configuration.

If we don't get a HTTP_VERSION header, just set it to the default
Transport HTTP version.

Closes #2256.
Closes #2310

Reviewed By: @fredemmott

Differential Revision: D1258381

Pulled By: @alexmalyshev