Add a match() function for use by IR optimizer tests
This avoids the cumbersome testing of each component of an instruction
to ensure it matches the expected output.

The one thing this can't quite accomplish is comparing ExtraData. The
root of the problem is that IRExtraData subclasses don't generally have
an equality function, and as far as I can tell you'd need to implement
one manually for each subclass.

You could imagine something even more wack like extending this mechanism
to entire Blocks. I haven't spent much time seeing how feasible that is.

Reviewed By: @edwinsmith

Differential Revision: D1257438