Remove ArrayData::endRef, add a few comments to ArrayIterator
Also some random things from auditing callsites of things
like getValueRef and secondRef.  I was trying to figure out if it was
possible to require that callers of these functions don't ever box the
Variant& that is returned, but it looks like this is not the case.
(Which means the hypothetical bug described in #4055855 is probably a
real bug.)

Reviewed By: @edwinsmith

Differential Revision: D1258707