Start jsonizing the framework test runner
This is the first of a few diffs that will begin to move to parsing a json format for test output instead of relying on a bunch of fragile regex to get our output.

First up: Have the test command create the temporary json files that will be needed when we squash together the results into one big file.

Reviewed By: @fredemmott

Differential Revision: D1259518