Clean up and document ArrayData function table a bit
Consolidate the shared functions in ArrayCommon, and document
the behavior of each entry point.  (ArrayCommon is pre-emptively a
struct so it can be friends of things, but so far it didn't actually
need any of that.)

Since all the virtuals are exposed as public functions I'll probably
move the documentation to the header at some point, but I didn't do
that yet.  (I feel like ArrayData really needs its own cleanup pass,
at which point we can pull the docs up there.)

There were at least two places where the array functions table was
wrong.  One is fixed in D1261567, the other looked older and only
fixed in this diff: AppendWithRef called the wrong function for
APCLocalArray, NameValueTableWrapper, and ProxyArray.

Reviewed By: @dariorussi

Differential Revision: D1261653