Hide generic ArrayData functions in MixedArray*'s
If you know you have a MixedArray*, generally you should be
making non-virtual calls, so using the instance methods on ArrayData
is not desirable (we can't use "final" to get this because they aren't
real virtuals).  The non-virtual calls take ArrayData*'s, so unless
you need to poke directly into the arrays (e.g. iter_next helpers)
generally you shouldn't even make a MixedArray* to call them.  In
light of that, it seems like we don't really need/want MixedArray* to
be implicitly convertible to ArrayData* outside of its
implementation---an explicit conversion is good enough and will avoid
accidentally using it as if it were a normal class.

Reviewed By: @swtaarrs

Differential Revision: D1262154