json will not work short-term because of phpunit < 4.0 vs >= 4.0 inconistencies
This reverts commit dd39a105cc87208fe6b8c509f6505fc26c67067f.

I thought I tried jsonizing the runner about 6 months ago and there was a reason it wouldn't work right....

https://github.com/facebook/hhvm/pull/2355

could be the reason why. Phpbb3, for example, includs parts of phpunit from *their* vendor directory (which is probably 3.7).

So I am backing this out for now. We would have to get all 5-6 frameworks to utilize 4.0+ in their code.

@fredemmott said maybe we could try adding a custom TestSuiteListener instead. I will try that.

 --printer <printer>       TestSuiteListener implementation to use.

Reviewed By: @ptarjan

Differential Revision: D1263494