Make translateRegion follow the arcs specified in the region
With bytecode control flow enabled, translateRegion was considering
all the arcs in the entire subgraph induced by the region blocks as
being part of the region. This diff changes translateRegion to
actually follow the arcs specified in the region.

This diff also adds support for regions containing multiple blocks
corresponding to the same SrcKey.

Reviewed By: @bertmaher

Differential Revision: D1256319