Fix iteration issues in ext_wddx
It wasn't handling KindOfRefs and it was using getChildren
instead of array iteration (so for e.g. if the array came from APC it
would do something else entirely).  (I'm unsure if it's supposed to
support iterable objects, but that's something else...)  Not handling
KindOfRefs seems to be deliberate (to avoid handling recursive arrays),
so that behavior is preserved for now.

Reviewed By: @scannell

Differential Revision: D1268081