Some refactoring in APCLocalArray and NameValueTable/$GLOBALS use in extensions
- Remove more implicit conversions to ArrayData from these two
    arrays

  - Remove calls to get_global_variables.  (We can probably remove it,
    and the way globals are initialized in the hhvm_main_stubs TARGETS
    without much more effort.)

  - Add new php-globals.h with documented functions for accessing
    global variables from extensions.  The API is not complete (no
    unset, exists, etc for now), but I think we deliberately should
    separate this API from the array API---$GLOBALS is much more
    dangerous...

On that note, I realized while looking at this that any use of things
like GetValueRef or LvalAt, where we are operating on an Array that's
coming in from user code (which could be $GLOBALS) is probably a bug.
Unless you are sure that you aren't going to run php code as a result
of whatever you are trying to do (e.g. if you are doing a nested set
and the old value runs a php destructor)....  It seems like even if
you don't store the lval pointer, it could result in things like
running code in MixedArray::SetStr while we've already released the
array we were operating on at the point that we decref an old value...
depending on the order of stores in those functions you could write to
freed memory.

Are all those APIs just fundamentally broken by NameValueTable not
supporting COW, or am I missing something?

Reviewed By: @jano

Differential Revision: D1269305