Delay using of runtime option EvalProfileHWEnable before it is parsed.
The diff is to delay using of the runtime option EvalProfileHWEnable
before it is parsed.

Currently, the value of RuntimeOption::EvalProfileHWEnable is used in
the constructor of HardwareCounterImpl, which is eventually called by
init_thread_locals inside execute_program in
program-function.cpp. However, runtime option parsing happens only
after init_thread_locals is called, which caused "using before
parsing" problem.

To solve the problem, move RuntimeOption::EvalProfileHWEnable usage
from the constructor of the HardwareCounterImpl to a seperate
initilization method. The method is then called when read() happens at
runtime. read() is expected to be called only after all options are
parsed.

Reviewed By: @ptarjan

Differential Revision: D1267669