Tracelet postconditions should be computed using the main trace
We record postconditions based on the types of stack/local
values at the end of translation, but nothing says that we have the
values from the main trace at that point.  The solution in this diff
is to find the main exit and explicitly pull its state out of
FrameState; making that work requires saving each block's FrameState
at the end of that block.

Reviewed By: @ottoni

Differential Revision: D1272977