Improve behavior when serializing unserializable builtin classes
Instead of serializing an object of some unrelated type,
serialize a null.  The old way had a dynamic property telling you
which type you messed up with, but this is not nearly as useful as a
stacktrace at serialize time, as the code doing the serialization is
generally going to be the code making the error.  "Null" is also going
to make this easier to deal with for private property inference---if
we've determined that a private property has any object type,
currently either it is going to be nullable or one of the collection
types (which are known to be serializable).  Presumably the old dummy
continuation method was also easier for private property inference
because dummy continuations derived from Continuation?

Reviewed By: @jano

Differential Revision: D1274152