Genericize Assert* opcodes so they can contain array information
Summary: Instead of AssertT{L,Stk}, AssertObj{L,Stk}, and then having
to add an AssertArr{L,Stk}, this restructures the assert opcodes so
the same RepoAuthType struct that we use to provide types for
properties can also be used in the bytecode stream.  It also adds
initial array type information through a repo-global "array type
table"; for now only packed-like arrays of known size are hooked up to
this, which also means modifying JIT::Type to be able to carry that
information around.

This diff also fixes the ability to guard on specialized array types
when there is an assertion from hhbbc that something is of type Array:
before this change it was breaking that code in Translator::analyze()
and we would never generate the specialized guards in this situation.
Specialized guards on specific collection types is still broken if
there is an assert of type Obj.

For now this diff removes the PredictT opcodes, which hopefully folks
are ok with---I think it will be easy to re-add them as
RepoAuthType-style things later (I could do that in this diff if
anyone feels strongly).

Reviewed By: @swtaarrs

Differential Revision: D1293633