Implement f_strpbrk using libc's strcspn when there's no null bytes
Seems to get pack the instructions regression from the change
to a manual loop.  It looks like the libc implementation of strcspn is
similar to the implementation of strpbrk (various machine feature
detection with different implementations).  I think it is also
possible to use strcspn even if the set has null bytes in it if you
remove the null byte and record whether it should count as a hit
(since strcspn will always stop on a null byte), but this doesn't do
that.

Reviewed By: @billf

Differential Revision: D1276594