Small fixes so hhbbc doesn't assert-fail some cases with malformed classes
I realized I was assuming there must be at least one
ClassInfo for the context class here, but that's not true if the class
is malformed enough that we decided it would always fatal to define
it.  This also means another always_assert about unique classes was
too strong (but earlier it was useful for debugging), so remove that.

Reviewed By: @jano

Differential Revision: D1274285