Fix crash when trying to change staticness of a property
If we found that a property had been redeclared as static
after originally being declared non-static, we hunted through the
parent preClasses to find the original definition. But the original
definition might have been in a trait. Also, the prop itself knows
where it was originally defined, so just use that.

Reviewed By: @ottoni

Differential Revision: D1275832