Allow NativeFuncCaller to handle up to 32 args regardless of type
Since we're hard-coding knowledge of the ABI into the JIT
anyway, might as well teach interp mode how to take advantage of it.

The AMD64/ARMv8 ABIs don't actually care about the specific order
of INT/SSE args with relation to one another up to the stack frontier.

Take advantage of this by reordering the first 8 double args to the start
of the function call and autoboxing all remaining doubles to ints so that
we only have to create fcall typedefs for arbitrary numbers of INT args.

For functions which take more than 8 doubles, but fewer than 6(or 8 on ARM)
INT args, we pad GPargs with nulls to land the extra SSEs on the stack.

This drops our complexity from O(2^N) to O(M*N)
M == Number of SIMD registers (8 for both amd64 and ARMv8)
N == Maximum number of args, type irrelevant (arbitrarily set to 32)

Build time (on my box) for native.cpp has dropped to ~15s

There are no longer any templates (recursive or otherwise) left in native.cpp

I expect this to be a slight perf regression for interp mode,
since we're pushing args into intermediate vectors,
but since basically every global function is now FCallBuiltin
capable, it shouldn't effect those in JIT mode.

JIT calls to HNI object methods may see a perf hit, but that's
something I'll be able to win back by making them eligible
for FCallBuiltin in a followup diff.

Reviewed By: @edwinsmith

Differential Revision: D1263316