Implement ConcatN opcode
Summary: Introduces a ConcatN opcode for when we concatenate a small, fixed
number of strings, as well as IR ConcatStrN instructions for each
supported value of n.

As is, only HHBBC generates these, and currently only when all operands
are known to be strings.  The heuristic for the rewrites is very basic,
but should catch some common use cases.

As part of this implementation, we add a stack of "sources" to
HHBBC::State, which tracks the opcode that produced each value on the
evaluation stack.

Reviewed By: @jdelong

Differential Revision: D1281404