Expand typedefs when doing a string concat
In strict mode, we'd `raise Ignore` because we thought this
class didn't exist, even though it was just an unexpanded typedef. I'm
not sure that `raise Ignore` is the right thing to do -- I guess we'll
have yelled elsewhere about an undefined class? -- but in any event it
was causing some code to not get checked and we even failed to verify
that the string concat was legal.

This is kind of obnoxious -- again, I hate that every code anywhere in
`hh_server` that has to look up a class has to be aware of typedefs.
This is the second or third bug where we've missed a spot where typedefs
should be expanded, leading to an inscruitable error message.

That said, I don't know of a better strategy offhand :\

Fixes #2272

Reviewed By: @elgenie

Differential Revision: D1254679