Remove the explicit Variant(const StaticString*) constructor
If you have explicit constructors that can't be used (because
it's not an explicit call), implicit constructors may be used after
other conversions.  So, for example:

  struct foo {
    foo(bool);
    explicit foo(const char*);
  };
  void f(foo);
  int main() {
    const char* a = "heh";
    f(a);   // calls foo::foo(bool)  :-(
  }

Which is error prone and surprising.  This was why constructing String
implicitly with const StringData*'s was calling Variant(bool) and then
String(Variant&&).  (Deleted another way in D1274141, which seems
worth leaving in also.)

Reviewed By: @jano

Differential Revision: D1281739