Disallow closing tag even in decl mode
HHVM disallows this and so we should be consistent. IMO it
makes sense for us to be stricter. Having a closing tag indicates a
fundamental misunderstanding of the way Hack works -- mixed HTML and PHP
just doesn't fly anymore. Hack is a programming language, not an HTML
processor. So it makes sense to ban this even in decl mode, IMO.

Closes #2390

Reviewed By: @pikatchu

Differential Revision: D1280962