Fix namespaced classes referring to classes in no namespace
`elaborate_id` wasn't idempotent; you could change a
fully-qualified identifier into an unqualified one, which would then get
fully-qualified into the wrong namespace on a subsequent pass.

Fix this by carrying through the fully-qualified namespace name
everywhere, which we probably should have done in the first place, and
then snipping it when printing out. Error messages then universally are
just missing the leading '\', which is nice and clean for FB and nice
and uniform for everyone else.

Actually doing this involved a lot of annoying tracking down of where we
print out a type name (everywhere) and where we pattern match against a
type name (everywhere). Probably missed a few spots, but the tests all
pass.

Fixes #2280

Reviewed By: @pikatchu

Differential Revision: D1254406